+++ packages/java/src/browser/java-frontend-module.ts
@@ -25,11 +25,7 @@ import { JavaCommandContribution } from './java-commands';
 import { JavaLabelProviderContribution } from './java-label-provider';
 import { JavaResourceResolver } from './java-resource';
 import { JavaEditorTextFocusContext } from './java-keybinding-contexts';
-import { JavaTextmateContribution } from './monaco-contribution/java-textmate-contribution';
-import { LanguageGrammarDefinitionContribution } from '@theia/monaco/lib/browser/textmate/textmate-contribution';
 import { bindJavaPreferences } from './java-preferences';
-
-import './monaco-contribution';
 
 export default new ContainerModule(bind => {
     bindJavaPreferences(bind);
@@ -48,6 +44,4 @@ export default new ContainerModule(bind => {
     bind(ResourceResolver).toService(JavaResourceResolver);
 
     bind(LabelProviderContribution).to(JavaLabelProviderContribution).inSingletonScope();
-
-    bind(LanguageGrammarDefinitionContribution).to(JavaTextmateContribution).inSingletonScope();
 });
