+++ packages/textmate-grammars/src/browser/textmate-grammars-frontend-module.ts
@@ -46,6 +46,7 @@ import { ShellContribution } from './shell';
 import { TclContribution } from './tcl';
 import { XmlContribution } from './xml';
 import { XslContribution } from './xsl';
+import { JavaContribution } from './java';
 import { LanguageGrammarDefinitionContribution } from '@theia/monaco/lib/browser/textmate/textmate-contribution';
 
 export default new ContainerModule(bind => {
@@ -141,4 +142,7 @@ export default new ContainerModule(bind => {
 
     bind(XslContribution).toSelf().inSingletonScope();
     bind(LanguageGrammarDefinitionContribution).toService(XslContribution);
+
+    bind(JavaContribution).toSelf().inSingletonScope();
+    bind(LanguageGrammarDefinitionContribution).toService(JavaContribution);
 });
